rollback のテストがエラーになっていたため, 別クラスでテストするよう修正
