+++ tests/Eccube/Tests/Service/PluginServiceTest.php
@@ -31,18 +31,13 @@ class PluginServiceTest extends AbstractServiceTestCase
 {
     protected $app;
 
-    public function setUp()
-    {
-        parent::setUp();
-    }
-
     /*
        正しいプラグインの条件
        * tar/zipアーカイブである
        * 展開した直下のディレクトリにconfig.ymlがあり、正しいymlファイルである
        * config.ymlの必須要素が規定の文字数、文字種で定義されている
        * event.ymlが存在する場合、正しいymlである
- 
+
      */
 
     // テスト用のダミープラグインを配置する
@@ -77,7 +72,7 @@ class PluginServiceTest extends AbstractServiceTestCase
 
         $tar = new \PharData($tmpfile);
         $tar->addFromString('config.yml',Yaml::dump($config));
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
 
         // インストールできるか
         $this->assertTrue($service->install($tmpfile));
@@ -107,7 +102,7 @@ class PluginServiceTest extends AbstractServiceTestCase
         $this->setExpectedException(
           '\Eccube\Exception\PluginException', 'config.yml not found or syntax error'
         );
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
 
         // インストールするプラグインを作成する
         $tmpname="dummy".sha1(mt_rand());
@@ -118,17 +113,17 @@ class PluginServiceTest extends AbstractServiceTestCase
         $tar->addFromString('dummy','dummy');
         // インストールできるか
         $service->install($tmpfile);
-        
+
     }
 
     // config.ymlのフォーマット確認
     public function testConfigYmlFormat()
     {
 #        self::markTestSkipped();
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
         $tmpname='dummy'.mt_rand();
         $tmpfile=sys_get_temp_dir().'/dummy'.mt_rand();
-        
+
 
         // 必須項目のチェック
         $config=array();
@@ -139,7 +134,7 @@ class PluginServiceTest extends AbstractServiceTestCase
             file_put_contents($tmpfile,Yaml::dump($config));
             $service->checkPluginArchiveContent($tmpfile);
             $this->fail("testConfigYmlFormat dont throw exception.");
-        }catch(\Eccube\Exception\PluginException $e){ } 
+        }catch(\Eccube\Exception\PluginException $e){ }
 
         $config=array();
         $config['name'] = $tmpname;
@@ -170,7 +165,7 @@ class PluginServiceTest extends AbstractServiceTestCase
             file_put_contents($tmpfile,Yaml::dump($config));
             $service->checkPluginArchiveContent($tmpfile);
             $this->fail("testConfigYmlFormat dont throw exception.");
-        }catch(\Eccube\Exception\PluginException $e){ } 
+        }catch(\Eccube\Exception\PluginException $e){ }
 
         $config=array();
         $config['name'] = $tmpname;
@@ -209,7 +204,7 @@ class PluginServiceTest extends AbstractServiceTestCase
     public function testnstallPluginMalformedConfigError()
     {
         #self::markTestSkipped();
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
         $tmpdir=$this->createTempDir();
         $tmpfile=$tmpdir.'/plugin.tar';
         $tar = new \PharData($tmpfile);
@@ -289,14 +284,14 @@ EOD;
         $event['eccube.event.app.after'][] = array("dummyHandlerLast",'LAST');
         $tar->addFromString('event.yml',Yaml::dump($event));
 
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
 
         // インストールできるか
         $this->assertTrue($service->install($tmpfile));
         $rep= $this->app['eccube.repository.plugin'];
 
         $plugin=$rep->findOneBy(array('code'=>$tmpname)); // EntityManagerの内部状態を一旦クリア // associationがうまく取れないため
-        $this->app['orm.em']->detach($plugin); 
+        $this->app['orm.em']->detach($plugin);
 
 
         // インストールした内容は正しいか
@@ -365,7 +360,7 @@ EOD;
         $this->assertEquals($plugin->getVersion(),$tmpname."u");
 
         // イベントハンドラが新しいevent.ymlと整合しているか(追加、削除)
-        $this->app['orm.em']->detach($plugin); 
+        $this->app['orm.em']->detach($plugin);
         $this->assertTrue((boolean)$plugin=$rep->findOneBy(array('code'=>$tmpname)));
         $this->assertEquals(3,count($plugin->getPluginEventHandlers()->toArray()));
 
@@ -385,7 +380,7 @@ EOD;
         }
         // 追加されたファイルが配置されているか
         $this->assertFileExists(__DIR__."/../../../../app/Plugin/$tmpname/update_dummy");
- 
+
         // アンインストールできるか
         $this->assertTrue($service->uninstall($plugin));
         // ちゃんとファイルが消えているか
@@ -395,51 +390,7 @@ EOD;
         $this->assertFileNotExists(__DIR__."/../../../../app/Plugin/$tmpname/DummyEvent.php");
     }
 
-
-     // インストーラが例外を上げた場合ロールバックできるか
-    public function testInstallPluginWithBrokenManager()
-    {
-        // インストールするプラグインを作成する
-        $tmpname="dummy".sha1(mt_rand());
-        $config=array();
-        $config['name'] = $tmpname;
-        $config['code'] = $tmpname;
-        $config['version'] = $tmpname;
-
-        $tmpdir=$this->createTempDir();
-        $tmpfile=$tmpdir.'/plugin.tar';
-
-        $tar = new \PharData($tmpfile);
-        $tar->addFromString('config.yml',Yaml::dump($config));
-        $dummyManager=<<<'EOD'
-<?php
-namespace Plugin\@@@@ ;
-
-use Eccube\Plugin\AbstractPluginManager;
-class PluginManager extends AbstractPluginManager
-{
-    public function install($plugin,$app)
-    {
-        throw new \Exception('hoge',1);
-    }
-
-}
-
-EOD;
-        $dummyManager=str_replace('@@@@',$tmpname,$dummyManager); // イベントクラス名はランダムなのでヒアドキュメントの@@@@部分を置換
-        $tar->addFromString("PluginManager.php" , $dummyManager);
-        $service = $this->app['eccube.service.plugin']; 
-        try{
-            $this->assertTrue($service->install($tmpfile));
-            $this->fail("BrokenManager dont throw exception.");
-        }catch(\Exception $e){ }
-
-        // インストーラで例外発生時にテーブルやファイスシステム上にゴミが残らないか
-        $this->assertFileNotExists(__DIR__."/../../../../app/Plugin/$tmpname");
-        $this->assertFalse((boolean)$plugin=$this->app['eccube.repository.plugin']->findOneBy(array('name'=>$tmpname)));
-    }
-
-     // インストーラが例外を上げた場合ロールバックできるか
+    // インストーラが例外を上げた場合ロールバックできるか
     public function testInstallPluginWithBrokenManagerAfterInstall()
     {
         // インストールするプラグインを作成する
@@ -462,7 +413,7 @@ use Eccube\Plugin\AbstractPluginManager;
 class PluginManager extends AbstractPluginManager
 {
     public function install($plugin,$app)
-    {    
+    {
         echo "";
     }
     public function uninstall($config,$app)
@@ -487,7 +438,7 @@ class PluginManager extends AbstractPluginManager
 EOD;
         $dummyManager=str_replace('@@@@',$tmpname,$dummyManager); // イベントクラス名はランダムなのでヒアドキュメントの@@@@部分を置換
         $tar->addFromString("PluginManager.php" , $dummyManager);
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
 
         // 正しくインストールでき、enableのハンドラが呼ばれないことを確認
         $this->assertTrue($service->install($tmpfile));
@@ -496,13 +447,13 @@ EOD;
         try{
             $this->assertTrue($service->enable($plugin));// enableにしようとするが、例外発生
         }catch(\Exception $e){ }
-        $this->app['orm.em']->detach($plugin); 
+        $this->app['orm.em']->detach($plugin);
         $this->assertTrue((boolean)$plugin=$this->app['eccube.repository.plugin']->findOneBy(array('name'=>$tmpname)));
         $this->assertEquals(Constant::DISABLED,$plugin->getEnable()); // プラグインがdisableのままになっていることを確認
 
     }
 
-     // インストーラを含むプラグインが正しくインストールできるか 
+     // インストーラを含むプラグインが正しくインストールできるか
     public function testInstallPluginWithManager()
     {
         #self::markTestSkipped();
@@ -556,7 +507,7 @@ class PluginManager extends AbstractPluginManager
 EOD;
         $dummyManager=str_replace('@@@@',$tmpname,$dummyManager); // イベントクラス名はランダムなのでヒアドキュメントの@@@@部分を置換
         $tar->addFromString("PluginManager.php" , $dummyManager);
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
 
         // インストールできるか、インストーラが呼ばれるか
         ob_start();
@@ -603,7 +554,7 @@ EOD;
 
         $tar = new \PharData($tmpfile);
         $tar->addFromString('config.yml',Yaml::dump($config));
-        $service = $this->app['eccube.service.plugin']; 
+        $service = $this->app['eccube.service.plugin'];
 
         // インストールできるか
         $this->assertTrue($service->install($tmpfile));
